refactor(health): split diagnostics service
