+++ scripts/audit_maintainability/baselines/route_srp.json
@@ -3,7 +3,7 @@
   "rule": "route_srp_violations",
   "description": "No-regression baseline for route files that still mix raw SQL, json!() shaping, and crate::services calls.",
   "captured_at": "2026-06-28",
-  "total_count": 11,
+  "total_count": 10,
   "files": {
     "src/server/routes/agents_crud.rs": {
       "count": 1
@@ -18,9 +18,6 @@
       "count": 1
     },
     "src/server/routes/github.rs": {
-      "count": 1
-    },
-    "src/server/routes/health_api.rs": {
       "count": 1
     },
     "src/server/routes/memory_api.rs": {
