+++ src/server/routes/mod.rs
@@ -1,4 +1,3 @@
-pub mod active_session_audit;
 pub mod agents;
 mod agents_crud;
 mod agents_setup;
