+++ src/services/mod.rs
@@ -37,6 +37,8 @@ pub mod escalation_settings;
 pub mod gemini;
 pub mod git;
 pub mod github_issue_creation;
+pub mod health_active_session_audit;
+pub mod health_diagnostics;
 pub mod issue_announcements;
 pub mod kanban;
 pub mod kanban_cards;
